Dispatch logic for Unfold3d mirroring that of Unfold2d
